+++ CHANGELOG.md
@@ -1930,3 +1930,10 @@ Alternately, the _Daily Diary of my Programming Dreams_.
   12. Added [fiddles/react/fiddle-0012-MultipleInputs](fiddles/react/fiddle-0012-MultipleInputs)
   13. Added [fiddles/react/fiddle-0013-StateLifting](fiddles/react/fiddle-0013-StateLifting)
   14. Added [fiddles/react/fiddle-0014-ChildrenProp](fiddles/react/fiddle-0014-ChildrenProp)
+
+
+### 247_jquery_fiddle_64 ~ [Issue 247](https://github.com/bradyhouse/house/issues/247)
+
+* **JQuery**
+
+  1.  Added [fiddles/jquery/fiddle-0064-BootstrapJumboTronSearchBox](fiddles/jquery/fiddle-0064-BootstrapJumboTronSearchBox)